userId改为Long类型

+++ redbook-common/src/main/java/com/itcast/context/UserContext.java
@@ -4,13 +4,13 @@ package com.itcast.context;
  * 用户线程变量
  */
 public class UserContext {
-    public static final ThreadLocal<Integer> userThreadLocal = new ThreadLocal<>();
+    public static final ThreadLocal<Long> userThreadLocal = new ThreadLocal<>();
 
-    public static void setUserId(Integer userId) {
+    public static void setUserId(Long userId) {
         userThreadLocal.set(userId);
     }
 
-    public static Integer getUserId() {
+    public static Long getUserId() {
         return userThreadLocal.get();
     }
