+++ redbook-service/redbook-service-message/src/main/java/com/itcast/session/Session.java
@@ -6,25 +6,25 @@ import java.util.concurrent.ConcurrentHashMap;
 
 public class Session {
 
-    private static final ConcurrentHashMap<Integer, Channel> userIdChannelMap = new ConcurrentHashMap<>();
+    private static final ConcurrentHashMap<Long, Channel> userIdChannelMap = new ConcurrentHashMap<>();
 
-    private static final ConcurrentHashMap<Channel, Integer> channelUserIdMap = new ConcurrentHashMap<>();
+    private static final ConcurrentHashMap<Channel, Long> channelUserIdMap = new ConcurrentHashMap<>();
 
-    public static void bind(Integer userId, Channel channel) {
+    public static void bind(Long userId, Channel channel) {
         userIdChannelMap.put(userId, channel);
         channelUserIdMap.put(channel, userId);
     }
 
-    public static void unbind(Integer userId, Channel channel) {
+    public static void unbind(Long userId, Channel channel) {
         userIdChannelMap.remove(userId);
         channelUserIdMap.remove(channel);
     }
 
-    public static Integer getUserId(Channel channel) {
+    public static Long getUserId(Channel channel) {
         return channelUserIdMap.get(channel);
     }
 
-    public static Channel getChannel(Integer userId) {
+    public static Channel getChannel(Long userId) {
         return userIdChannelMap.get(userId);
     }
 }
