+++ redbook-service/redbook-service-order/src/main/java/com/itcast/config/MybatisPlusConfig.java
@@ -22,7 +22,7 @@ public class MybatisPlusConfig {
     @Bean
     public RequestInterceptor userIdRequestInterceptor() {
         return template -> {
-            Integer userId = UserContext.getUserId();
+            Long userId = UserContext.getUserId();
             if (userId != null) {
                 template.header("userId", String.valueOf(userId));
             }
