+++ redbook-service/redbook-service-search/src/main/java/com/itcast/model/pojo/History.java
@@ -27,6 +27,5 @@ public class History implements Serializable {
      * 用户id
      */
     @TableField("user_id")
-    private Integer userId;
+    private Long userId;
 }
-
